requirements.txt
certifi==2019.3.9
Django==2.1.7
django-markdownx==2.0.28
django-widget-tweaks==1.4.3
Markdown==3.1
Pillow==5.4.1
pytz==2018.9
sorl-thumbnail==12.5.0
